requirements/dev.txt
# pip install -r ./requirements/dev.txt

tensorflow==2.8.0
pandas==1.3.5
numpy==1.23.5
Pillow==9.3.0
tensorflow-hub==0.12.0
matplotlib==3.6.3
opencv-python==4.5.4.60
moviepy==1.0.3
wget==3.2
seaborn==0.12.2
scipy==1.10.0
scikit-learn==1.2.1

-f https://download.pytorch.org/whl/cu113/torch_stable.html

torch==1.11.0+cu113 # 1.10.1+cu111
torchaudio==0.11.0+cu113 # 0.10.1+cu111
torchvision==0.12.0+cu113  # 0.11.1+cu111